requirements.txt
python-docx>=1.1.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-ai-projects>=1.0.0b3
azure-identity>=1.15.0
openai>=1.30.0
semantic-kernel>=1.0.0
requests>=2.31.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
flake8>=7.0.0
aiohttp>=3.9.0
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
botframework-connector>=4.15.0
lxml>=5.0.0
